Refactor models and BLoC for readings, rooms, and tenants
Refactored Reading, Room, and Tenant models to use nullable IDs and updated types for rent and date fields. Introduced ReadingBloc with event and state management for readings. Updated UI and BLoC logic in readings, rooms, and tenants pages to use the new models and BLoC structure. Adjusted service and event signatures to match new model definitions and improved null safety throughout the codebase.

+++ lib/services/room_service.dart
@@ -16,7 +16,7 @@ class RoomService {
     }
   }
 
-  Future<Room> createRoom(String name, double rent) async {
+  Future<Room> createRoom(String name, int rent) async {
     final response = await http.post(
       Uri.parse(baseUrl),
       headers: {'Content-Type': 'application/json'},
@@ -30,7 +30,7 @@ class RoomService {
     }
   }
 
-  Future<Room> updateRoom(int id, String name, double rent) async {
+  Future<Room> updateRoom(int id, String name, int rent) async {
     final response = await http.put(
       Uri.parse('$baseUrl/$id'),
       headers: {'Content-Type': 'application/json'},
